Move Data and update_cumulative to the contract module
